fixed the jest config, package,json, refactored app.test.tsx,moved useDashboard to src/api/dashboard.ts
