+++ package.json
@@ -80,7 +80,7 @@
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx,json}": [
       "eslint --fix --max-warnings 0",
-      "yarn test",
+      "yarn test --findRelatedTests",
       "yarn prettier",
       "yarn lint"
     ],
